add event.rs
